UI
Сделал родительский класс абстрактным, перенёс в него логика отображения передвигаемого блока
